+++ lib/sbmt/pact/matchers/v1/equality.rb
@@ -5,8 +5,12 @@ module Sbmt
     module Matchers
       module V1
         class Equality < Sbmt::Pact::Matchers::Base
-          def initialize(serializer, value)
-            super(spec_version: Sbmt::Pact::Matchers::PACT_SPEC_V1, serializer: serializer, kind: "equalTo", value: value)
+          def initialize(template)
+            super(spec_version: Sbmt::Pact::Matchers::PACT_SPEC_V1, kind: "equality", template: template)
+          end
+
+          def as_plugin
+            "matching(equalTo, #{format_primitive(@template)})"
           end
         end
       end