+++ lib/sbmt/pact/matchers/v2/type.rb
@@ -5,10 +5,10 @@ module Sbmt
     module Matchers
       module V2
         class Type < Sbmt::Pact::Matchers::Base
-          def initialize(serializer, value)
-            raise MatcherInitializationError, "#{self.class}: value is not a primitive" unless value.is_a?(TrueClass) || value.is_a?(FalseClass) || value.is_a?(Numeric) || value.is_a?(String)
+          def initialize(template)
+            raise MatcherInitializationError, "#{self.class}: template is not a primitive" unless template.is_a?(TrueClass) || template.is_a?(FalseClass) || template.is_a?(Numeric) || template.is_a?(String)
 
-            super(spec_version: Sbmt::Pact::Matchers::PACT_SPEC_V2, serializer: serializer, kind: "type", value: value)
+            super(spec_version: Sbmt::Pact::Matchers::PACT_SPEC_V2, kind: "type", template: template)
           end
         end
       end