+++ spec/internal/pkg/client/pet_store/grpc/pet_store_services_pb.rb
@@ -10,6 +10,7 @@ module PetStore
       module V1
         module Pets
           class Service
+
             include ::GRPC::GenericService
 
             self.marshal_class_method = :encode